Debian tree

+++ gcc/testsuite/gcc.target/aarch64/aapcs64/test_18.c
@@ -9,10 +9,10 @@
 
 struct y
 {
-  long long p;
-  long long q;
-  long long r;
-  long long s;
+  long p;
+  long q;
+  long r;
+  long s;
 } v = { 1, 2, 3, 4 };
 
 struct z
